Renames BuilderatorRenderer to Builderator and Builderator to BuilderatorFacade.
